# swagger-api-documentation.yaml
openapi: 3.0.3
info:
  title: Agentic Itinerary Planner API
  description: |
    Comprehensive API for AI-powered travel itinerary planning and management.
    
    This API provides endpoints for:
    - Creating and managing travel itineraries
    - Real-time agent progress tracking via Server-Sent Events
    - Travel planning tools (weather, packing lists, photo spots, etc.)
    - Booking and payment processing
    - Health monitoring and testing
    
    All endpoints support CORS and work without authentication in development mode.
  version: 1.0.0
  contact:
    name: Agentic Itinerary Planner Team
    email: support@agenticitinerary.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.agenticitinerary.com/api/v1
    description: Production server
...
paths:
...
  /agents/stream:
    get:
      tags:
        - Agents
      summary: Stream agent events (SSE)
      description: |
        Establishes a Server-Sent Events connection to receive real-time updates
        about AI agent progress for itinerary generation.
      parameters:
        - name: itineraryId
          in: query
          required: true
          description: Itinerary ID to track
          schema:
            type: string
            example: "3"
      responses:
        '200':
          description: SSE connection established
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: connected
                  data: Connected to agent stream for itinerary: 3
                  
                  event: agent-progress
                  data: {"kind":"planner","status":"running","progress":45,"message":"Generating activities..."}
  /agents/events/{itineraryId}:
    get:
      tags:
        - Agents
      summary: Stream agent events by ID (SSE)
      description: |
        Alternative SSE endpoint for streaming agent events for a specific itinerary.
        This endpoint provides the same functionality as /agents/stream but uses
        path parameter instead of query parameter.
      parameters:
        - name: itineraryId
          in: path
          required: true
          description: Itinerary ID to track
          schema:
            type: string
            example: "3"
      responses:
        '200':
          description: SSE connection established
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  event: connected
                  data: Connected to agent stream for itinerary: 3
                  
                  event: agent-progress
                  data: {"kind":"planner","status":"running","progress":45,"message":"Generating activities..."}
  /agents/{itineraryId}/status:
    get:
      tags:
        - Agents
      summary: Get agent status
      description: Get current status of agents for a specific itinerary (polling alternative to SSE)
      parameters:
        - name: itineraryId
          in: path
          required: true
          description: Itinerary ID
          schema:
            type: string
            example: "3"
      responses:
        '200':
          description: Agent status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatusResponse'

  # Tools Endpoints
...
  /payments/razorpay/order:
    post:
      tags:
        - Bookings
      summary: Create Razorpay order
      description: Creates a new Razorpay payment order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RazorpayOrderReq'
      responses:
        '200':
          description: Razorpay order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RazorpayOrderRes'
  /payments/razorpay/webhook:
    post:
      tags:
        - Bookings
      summary: Razorpay webhook
      description: Handles Razorpay payment webhook notifications
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              description: Raw webhook payload
      responses:
        '200':
          description: Webhook processed successfully
  /providers/{vertical}/{provider}:book:
    post:
      tags:
        - Bookings
      summary: Execute provider booking
      description: Executes booking with a specific provider after payment confirmation
      parameters:
        - name: vertical
          in: path
          required: true
          description: Booking vertical (hotels, flights, activities, etc.)
          schema:
            type: string
            example: "hotels"
        - name: provider
          in: path
          required: true
          description: Provider name
          schema:
            type: string
            example: "booking"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderBookReq'
      responses:
        '200':
          description: Provider booking executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingRes'
  /bookings:
    get:
      tags:
        - Bookings
      summary: Get user bookings
      description: Retrieves all bookings for the current user
      parameters:
        - name: page
          in: query
          required: false
          description: Page number (0-based)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: Page size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Bookings retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookingRes'
  /bookings/{bookingId}:
    get:
      tags:
        - Bookings
      summary: Get booking by ID
      description: Retrieves a specific booking by its ID
      parameters:
        - name: bookingId
          in: path
          required: true
          description: Booking ID
          schema:
            type: string
            example: "123"
      responses:
        '200':
          description: Booking retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingRes'
  /bookings/{bookingId}:cancel:
    post:
      tags:
        - Bookings
      summary: Cancel booking
      description: Cancels a specific booking
      parameters:
        - name: bookingId
          in: path
          required: true
          description: Booking ID
          schema:
            type: string
            example: "123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelBookingReq'
      responses:
        '200':
          description: Booking canceled successfully

  # Test Endpoints
...
components:
  schemas:
...
    AgentStatusResponse:
      type: object
      properties:
        itineraryId:
          type: string
          description: Itinerary ID
          example: "3"
        status:
          type: string
          description: Agent status
          example: "active"
        message:
          type: string
          description: Status message
          example: "Processing itinerary..."

    # Tools DTOs
...
    RazorpayOrderReq:
      type: object
      required:
        - itemType
        - itineraryId
        - amount
        - currency
      properties:
        itemType:
          type: string
          description: Type of item being booked
          example: "hotel"
        itineraryId:
          type: string
          description: Associated itinerary ID
          example: "3"
        amount:
          type: integer
          format: int64
          minimum: 1
          description: Amount in paise (smallest currency unit)
          example: 250000
        currency:
          type: string
          description: Currency code
          example: "INR"
        meta:
          type: object
          additionalProperties: true
          description: Additional metadata
    RazorpayOrderRes:
      type: object
      properties:
        orderId:
          type: string
          description: Razorpay order ID
          example: "order_123456789"
        amount:
          type: integer
          format: int64
          description: Order amount in paise
          example: 250000
        currency:
          type: string
          description: Currency code
          example: "INR"
        receipt:
          type: string
          description: Order receipt
          example: "receipt_123"
    ProviderBookReq:
      type: object
      required:
        - payment
        - item
        - itineraryId
      properties:
        payment:
          $ref: '#/components/schemas/PaymentDetails'
        item:
          $ref: '#/components/schemas/ItemDetails'
        itineraryId:
          type: string
          description: Associated itinerary ID
          example: "3"
    PaymentDetails:
      type: object
      required:
        - orderId
        - paymentId
        - signature
      properties:
        orderId:
          type: string
          description: Razorpay order ID
          example: "order_123456789"
        paymentId:
          type: string
          description: Razorpay payment ID
          example: "pay_123456789"
        signature:
          type: string
          description: Payment signature
          example: "signature_123"
    ItemDetails:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Item token
          example: "item_token_123"
        details:
          type: object
          additionalProperties: true
          description: Item-specific details
    BookingRes:
      type: object
      properties:
        bookingId:
          type: string
          description: Booking ID
          example: "123"
        status:
          type: string
          description: Booking status
          example: "confirmed"
        providerConfirmationId:
          type: string
          description: Provider confirmation ID
          example: "provider_123"
        itineraryId:
          type: string
          description: Associated itinerary ID
          example: "3"
        createdAt:
          type: string
          format: date-time
          description: Booking creation timestamp
        details:
          type: object
          additionalProperties: true
          description: Booking details
    CancelBookingReq:
      type: object
      properties:
        reason:
          type: string
          description: Cancellation reason
          example: "Change of plans"
        meta:
          type: object
          additionalProperties: true
          description: Additional metadata

    # Error DTOs
...
  securitySchemes:
    # No authentication required in development mode
    # This can be extended for production use
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (not currently required)
security:
  # No security requirements in development mode
  - {}